+++ TradinatorCore/Public/Indicators/WMA.h
@@ -15,14 +15,14 @@ class WMA : public Indicator
 public:
 	WMA() : Indicator() {}
 	WMA(EIndicatorSource source) : Indicator(source) {}
-	WMA(size_t length) : Indicator(length) {}
-	WMA(EIndicatorSource source, size_t length) : Indicator(source, length) {}
-	WMA(size_t length, std::weak_ptr<Security> security) : Indicator(length, security) {}
-	WMA(EIndicatorSource source, size_t length, std::weak_ptr<Security> security) : Indicator(source, length, security) {}
+	WMA(uint64_t length) : Indicator(length) {}
+	WMA(EIndicatorSource source, uint64_t length) : Indicator(source, length) {}
+	WMA(uint64_t length, std::weak_ptr<Security> security) : Indicator(length, security) {}
+	WMA(EIndicatorSource source, uint64_t length, std::weak_ptr<Security> security) : Indicator(source, length, security) {}
 
 
 	virtual std::vector<std::vector<double>> Calculate() override;
-	void CalculateRaw(const double* input, double* output, int64_t window_size, int64_t data_size);
+	void CalculateRaw(const double* input, double* output, uint64_t window_size, uint64_t data_size);
 
 	virtual std::string GetName() const override { return std::format("Weighted Moving Average ({})", m_length); }
 	virtual EIndicatorType IndicatorType() const override { return EIndicatorType::E_WMA; }
